Code style,comments,seperate
